Protection against 0 length subheaders
When we actually populate the subHeaders list we keep track of which subheaders have length zero. By using the size of that list we can achieve the same expected result; however, can now better support zero length subheaders. The +12 at the end might not be accurate for Java (I've ported all the code to D); however, I've confirmed with what I'm working on, adding the additional length of a subheader to this was necessary.